Include unit tests with tests() in test.py and refactored
