+++ config.py
@@ -6,11 +6,13 @@ credentials_dir='/root/credentials.csv'
 base_url='https://nfc.ai/'
 mcp_uri='mcp/'
 api_url=f'{base_url}{mcp_uri}'
+map_url=f'{base_url}map/'
 current_forecasts_api=f'{api_url}forecasts'
 live_storms_api=f'{api_url}live-storms'
 forecasts_live_table = 'forecasts_live'
 forecasts_archive_table = 'forecasts_archive'
 daily_ingest_sql_path = 'docs/daily_ingest.sql'
 daily_prompt_path = 'docs/daily_report_ingest_prompt.txt'
-wp_create_post_url='https://fluids.ai/wp-json/wp/v2/posts'
+fluids_url = 'https://fluids.ai/'
+wp_create_post_url=f'{fluids_url}wp-json/wp/v2/posts'
 gpt_model = '/data/phi-4.gguf'
